// Backend/src/enums/complaintAgainst.ts
export enum ComplaintAgainst {
    ACCOUNTANTGENERALSINDH = 'Accountant General Sindh',
    AGRICULTURESUPPLYANDPRICESDEPARTMENT = 'Agriculture Supply and Prices Department',
    AUQAFRELIGIOUSAFFAIRSZAKATANDUSHRDEPARTMENT = 'Auqaf Religious Affairs Zakat and Ushr Department',
    BENAZIRBHUTTOSHAHEEDHUMANRESOURCERESEARCHANDDEVELOPMENTBOARD = 'Benazir Bhutto Shaheed Human Resource Research and Development Board',
    COOPERATIONDEPARTMENT = 'Cooperation Department',
    CULTURETOURISMANDANTIQUITIESDEPARTMENT = 'Culture Tourism and Antiquities Department',
    EDUCATIONANDLITERACYDEPARTMENT_COLLEGESEDUCATION = 'Education and Literacy Department (Colleges Education)',
    EDUCATIONANDLITERACYDEPARTMENT_SCHOOLSEDUCATION = 'Education and Literacy Department (Schools Education)',
    ENERGYDEPARTMENT = 'Energy Department',
    ENQUIRIESANDANTICORRUPTIONESTABLISHMENT = 'Enquiries and Anti Corruption Establishment',
    ENVIRONMENTCLIMATECHANGEANDCOSTALDEVELOPMENTDEPARTMENT = 'Environment Climate Change and Costal Development Department',
    ENVIRONMENTALALTERNATEENERGY = 'Environmental Alternate Energy',
    EXCISETAXATIONANDNARCOTICSDEPARTMENT = 'Excise Taxation and Narcotics Department',
    FINANCEDEPARTMENT = 'Finance Department',
    FORESTANDWILDLIFEDEPARTMENT = 'Forest and Wildlife Department',
    FOODDEPARTMENT = 'Food Department',
    HEALTHDEPARTMENT = 'Health Department',
    HOMEDEPARTMENT = 'Home Department',
    HUMANRIGHTSDEPARTMENT = 'Human Rights Department',
    INDUSTRIESANDCOMMERCEDEPARTMENT = 'Industries and Commerce Department',
    INFORMATIONANDARCHIVESDEPARTMENT = 'Information and Archives Department',
    INFORMATIONSCIENCEANDTECHNOLOGYDEPARTMENT = 'Information Science and Technology Department',
    IRRIGATIONDEPARTMENT = 'Irrigation Department',
    KARACHIMUNICIPALCORPORATION = 'Karachi Municipal Corporation',
    KARACHIDEVELOPMENTAUTHORITY = 'Karachi Development Authority',
    KARACHIWATERANDSEWERAGEBOARD = 'Karachi Water and Sewerage Board',
    KATCHIABADIESANDSPATIALDEVELOPMENT = 'Katchi Abadies and Spatial Development',
    LABOURANDHUMANRESOURCESDEPARTMENT = 'Labour and Human Resources Department',
    LAWANDPARLIAMENTARYAFFAIRSDEPARTMENT = 'Law and Parliamentary Affairs Department',
    LIVESTOCKANDFISHERIESDEPARTMENT = 'Live Stock and Fisheries Department',
    LOCALGOVERNMENTANDHOUSINGTOWNPLANNINGDEPARTMENT = 'Local Government and Housing Town Planning Department',
    LYARIDEVELOPMENTAUTHORITY = 'Lyari Development Authority',
    MALIDEVELOPMENTAUTHORITY = 'Mali Development Authority',
    MINESANDMINERALDEVELOPMENTDEPARTMENT = 'Mines and Mineral Development Department',
    MINORITIESAFFAIRSDEPARTMENT = 'Minorities Affairs Department',
    PLANNINGANDDEVELOPMENTDEPARTMENT = 'Planning and Development Department',
    POLICEDEPARTMENT = 'Police Department',
    POPULATIONANDWELFAREDEPARTMENT = 'Population and Welfare Department',
    POWERDEPARTMENT = 'Power Department',
    PUBLICHEALTHENGINEERINGANDRURALDEVELOPMENTDEPARTMENT = 'Public Health Engineering and Rural Development Department',
    REVENUEDEPARTMENT = 'Revenue Department',
    SERVICESGENERALADMINISTRATIONANDCOORDINATIONDEPARTMENT = 'Services General Administration and Coordination Department',
    SINDHBUILDINGCONTROLAUTHORITY = 'Sindh Building Control Authority',
    SINDHEMPLOYESSOCIALSECURITYINSTITUTION = 'Sindh Employees Social Security Institution',
    SINDHHIGHEREDUCATIONCOMMISSIONKARACHI = 'Sindh Higher Education Commission Karachi',
    SINDHPUBLICSERVICECOMMISSION = 'Sindh Public Service Commission',
    SOCIALWELFAREDEPARTMENT = 'Social Welfare Department',
    SPECIALEDUCATIONDEPARTMENT = 'Special Education Department',
    SPORTANDYOUTHAFFAIRSDEPARTMENT = 'Sport and Youth Affairs Department',
    TRANSPORTANDMASSTRANSITDEPARTMENT = 'Transport and Mass Transit Department',
    WOMENDEVELOPMENTDEPARTMENT = 'Women Development Department',
    WORKANDSERVICESDEPARTMENT = 'Work and Services Department'
}