// Backend/src/enums/complaintOffice.ts
export enum ComplaintOffice {
    KARACHICENTRAL = 'Karachi (Central)',
    KARACHIEAST = 'Karachi (East)',
    KARACHISOUTH = 'Karachi (South)',
    THATTA = 'Thatta',
    SUKKUR = 'Sukkur',
    HYDERABAD = 'Hyderabad',
    BADIN = 'Badin',
    NAUSHAHROFEROZE = 'Naushahro Feroze',
    LARKANA = 'Larkana',
    MIRPURKHAS = 'Mirpurkhas',
    MITHI = 'Mithi',
    JACOBABAD = 'Jacobabad',
    DADU = 'Dadu',
    NAWABSHAH = 'Nawabshah',
    KHAIRPUR = 'Khairpur',
    GHOTKI = 'Ghotki'
}